Add playback control and some test files.
I added playback control, and test song files.
The test file is using a modified version of SongStub class to allow the creation of AudioInputStream.
Declaration of test file is in queueViewController.
Add an extran test line at line 107 in BottomViewController in order to get it to work with the modified SongStub.