Import file (#50)
* Updated ImportFile in Library to have caller pass song properties.
Updated FileImportTask to take song properties instead of trying to get them from the song file, which was proven to not work. Thanks a lot Oracle.
Updated Song to reflect that duration will now be stored in seconds, instead of microseconds.
Updated LocalLibrary to support changes to Library interface.
Updated DataManager to back changes to LocalLibrary.

* Updated ImportFile in Library to have caller pass song properties.
Updated FileImportTask to take song properties instead of trying to get them from the song file, which was proven to not work. Thanks a lot Oracle.
Updated Song to reflect that duration will now be stored in seconds, instead of microseconds.
Updated LocalLibrary to support changes to Library interface.
Updated DataManager to back changes to LocalLibrary.

* Changed Song ID system to use CRC-32 checksum of song file.

This'll provide globally unique IDs for each song that will only collide if two songs have the exact same song file.

* Updated LibraryStub to conform to new Library interface.

* Add UI....Need to do some listener

* Add File importer UI and connect it to back end

* Apply theme

* Minor alteration to logging data.

* Removed redundant logging output.

* Fixed importing with long IDs.

* Declare array

* Add song imports feature